Feat: base message versioning (#909)
* Add base message version to request

* Adjust tests

* Pass message version to the recipient

* Adjust clients and their tests

+++ packages/contracts-core/test/harnesses/libs/BaseMessageHarness.t.sol
@@ -43,7 +43,7 @@ contract BaseMessageHarness {
     }
 
     /// @notice Returns baseMessage's request field
-    function request(bytes memory payload) public pure returns (uint160) {
+    function request(bytes memory payload) public pure returns (uint192) {
         return Request.unwrap(payload.castToBaseMessage().request());
     }
